move #search into its own file
